+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Bug fixes:
 
 * Fix `Dir.glob` returning blank string entry with leading `**/` in glob and `base:` argument (@rwstauner).
 * Fix class lookup after an object's class has been replaced by `IO#reopen` (@itarato, @eregon).
+* Fix `Marshal.load` and raise `ArgumentError` when dump is broken and is too short (#3108, @andrykonchin).
 
 Compatibility: